pyproject.toml
[tool.poetry]
name = "wirikiki"
version = "1.0.0"
description = "A tiny desktop wiki"
homepage = "http://github.com/fdev31/wirikiki/"
authors = ["fdev31 <fdev31@gmail.com>"]
include = [
    "apps/**",
    "wirikiki/config/settings.toml",
    "wirikiki/web/**",
    "wirikiki/myKB/**",
]
packages = [
    { include = "wirikiki" },
]

[tool.poetry.scripts]
wirikiki = 'wirikiki.cli.wirikiki:run'
wirikiki-pwgen = 'wirikiki.cli.wirikiki_pwgen:run'

[tool.poetry.dependencies]
python = "^3.10"
python-jose = "^3.3.0"
tomli = "^2.0.1"
aiofiles = "^0.8.0"
orjson = "^3.7.11"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-multipart = "^0.0.5"
setproctitle = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
tox-poetry-installer = {extras = ["poetry"], version = "^0.8.4"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
